backport of commit be84613dc359421e470ca069b9592bc03ec1dbef
